Enhanced `PresentationManager.getElementProperties` to accept a custom content parser
