+++ common/api/presentation-common.api.md
@@ -124,6 +124,9 @@ export interface BooleanRulesetVariableJSON extends RulesetVariableBaseJSON {
     value: boolean;
 }
 
+// @internal (undocumented)
+export const buildElementProperties: (descriptor: Descriptor, item: Item) => ElementProperties;
+
 // @public
 export interface CalculatedPropertiesSpecification {
     categoryId?: string | CategoryIdentifier;
@@ -2766,8 +2769,6 @@ export class RpcRequestsHandler {
     // (undocumented)
     getDisplayLabelDefinition(options: DisplayLabelRequestOptions<IModelRpcProps, InstanceKey> & ClientDiagnosticsAttribute): Promise<LabelDefinition>;
     // (undocumented)
-    getElementProperties(options: SingleElementPropertiesRequestOptions<IModelRpcProps> & ClientDiagnosticsAttribute): Promise<ElementProperties | undefined>;
-    // (undocumented)
     getFilteredNodePaths(options: FilterByTextHierarchyRequestOptions<IModelRpcProps, RulesetVariableJSON> & ClientDiagnosticsAttribute): Promise<NodePathElementJSON[]>;
     // (undocumented)
     getNodePaths(options: FilterByInstancePathsHierarchyRequestOptions<IModelRpcProps, RulesetVariableJSON> & ClientDiagnosticsAttribute): Promise<NodePathElementJSON[]>;
@@ -2996,12 +2997,13 @@ export interface SetRulesetVariableParams<TVariable> extends CommonIpcParams {
 }
 
 // @public
-export interface SingleElementPropertiesRequestOptions<TIModel> extends RequestOptions<TIModel> {
+export interface SingleElementPropertiesRequestOptions<TIModel, TParsedContent = ElementProperties> extends RequestOptions<TIModel> {
+    contentParser?: (descriptor: Descriptor, item: Item) => TParsedContent;
     elementId: Id64String;
 }
 
 // @public
-export type SingleElementPropertiesRpcRequestOptions = PresentationRpcRequestOptions<SingleElementPropertiesRequestOptions<never>>;
+export type SingleElementPropertiesRpcRequestOptions = PresentationRpcRequestOptions<Omit_2<SingleElementPropertiesRequestOptions<never, never>, "contentParser">>;
 
 // @public
 export interface SingleSchemaClassSpecification {
