+++ common/changes/@itwin/presentation-common/presentation-content-parser-for-single-element-properties_2024-09-12-07-38.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/presentation-common",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/presentation-common"
+}
