+++ common/changes/@itwin/presentation-frontend/presentation-content-parser-for-single-element-properties_2024-09-12-07-38.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@itwin/presentation-frontend",
+      "comment": "Enhanced `PresentationManager.getElementProperties` to accept a custom content parser.",
+      "type": "none"
+    }
+  ],
+  "packageName": "@itwin/presentation-frontend"
+}
