+++ docs/changehistory/NextVersion.md
@@ -53,3 +53,14 @@ Approximation using `options.sampleMethod = EllipticalArcSampleMethod.AdaptiveSu
 Pictured below are triangulations of a DTM dataset with skirt points. At top is the result using default tolerance. Due to the skirt points having xy-distance greater than the default tolerance from actual terrain sites, they are included in the triangulation, resulting in undesirable near-vertical facets. At bottom is the result using `options.chordTol = 0.002`, which is sufficiently large to remove these artifacts:
 
 ![Toleranced Triangulations](./assets/triangulate-points-tolerance.jpg "Toleranced Triangulations")
+
+## Presentation
+
+### Custom content parser for creating element properties
+
+The `getElementProperties` function on the backend [PresentationManager]($presentation-backend) has two overloads:
+
+- For single element case, taking `elementId` and returning an data structure in the form of `ElementProperties`.
+- For multiple elements case, taking an optional list of `elementClasses` and returning properties of those elements. While the default form of the returned data structure is `ElementProperties`, just like in single element case, the overload allows for a custom parser function to be provided. In that case the parser function determines the form of the returned data structure.
+
+In this release the overload for single element case was enhanced to also take an optional custom content parser to make the two overloads consistent in this regard. In addition, the `getElementProperties` method on the frontend [PresentationManager]($presentation-frontend) has also been enhanced with this new feature to be consistent with the similar method on the backend.
