+++ presentation/common/src/presentation-common.ts
@@ -34,7 +34,6 @@ export * from "./presentation-common/KoqPropertyValueFormatter";
  */
 export * from "./presentation-common/PresentationRpcInterface";
 export * from "./presentation-common/RpcRequestsHandler";
-export * from "./presentation-common/ElementProperties";
 
 /**
  * @module UnifiedSelection
@@ -63,6 +62,7 @@ export * from "./presentation-common/content/TypeDescription";
 export * from "./presentation-common/content/Value";
 export * from "./presentation-common/content/ContentTraverser";
 export * from "./presentation-common/content/PropertyValueFormatter";
+export * from "./presentation-common/ElementProperties";
 
 /**
  * @module Hierarchies
